refactor: extract keybundle package from datastore (DRAFT — reverted)
Move watcher, bundle parsing, metrics, and lifecycle management into
central/signatureintegration/keybundle/ to separate background
processes from the CRUD datastore.

This commit is immediately reverted — kept only as a reference for PR
discussion about whether the extraction is worthwhile.
