+++ central/signatureintegration/datastore/singleton_test.go
@@ -5,7 +5,6 @@ import (
 
 	storeMocks "github.com/stackrox/rox/central/signatureintegration/store/mocks"
 	"github.com/stackrox/rox/pkg/signatures"
-	"github.com/stretchr/testify/assert"
 	"go.uber.org/mock/gomock"
 )
 
@@ -18,20 +17,6 @@ func TestSeedFirstInstall(t *testing.T) {
 	mockStore.EXPECT().Upsert(gomock.Any(), signatures.DefaultRedHatSignatureIntegration).Return(nil).Times(1)
 
 	seedRedHatDefaultSignatureIntegration(mockStore)
-}
-
-func TestStartKeyBundleWatcherDisabled(t *testing.T) {
-	t.Setenv("ROX_REDHAT_SIGNING_KEY_WATCH_INTERVAL", "0")
-
-	ctrl := gomock.NewController(t)
-	mockStore := storeMocks.NewMockSignatureIntegrationStore(ctrl)
-
-	old := bundleWatcher
-	defer func() { bundleWatcher = old }()
-	bundleWatcher = nil
-
-	startKeyBundleWatcher(mockStore)
-	assert.Nil(t, bundleWatcher)
 }
 
 func TestSeedSubsequentStartup(t *testing.T) {
