+++ central/signatureintegration/keybundle/bundle_test.go
@@ -1,4 +1,4 @@
-package datastore
+package keybundle
 
 import (
 	"testing"
