+++ central/signatureintegration/keybundle/log.go
@@ -1,0 +1,5 @@
+package keybundle
+
+import "github.com/stackrox/rox/pkg/logging"
+
+var log = logging.LoggerForModule()
